Improving precompile tasks
